+++ dart_sample/lib/sample_routes.dart
@@ -24,6 +24,7 @@ class SampleRoutes {
   static const String materialButtons = '/material-buttons';
   static const String checkbox = '/checkbox';
   static const String switchDemo = '/switch';
+  static const String radio = '/radio';
   static const String appBarLeadingWidth = '/appbar-leading-width';
   static const String appBarActionsPadding = '/appbar-actions-padding';
   static const String appBarIconTheme = '/appbar-icon-theme';